@@ -4,8 +4,6 @@ import M from "materialize-css";
 import { Timeline, Icon, Button } from "rsuite";
 import "./Home.css";
 
-// const profilePics = require("../images/emojis");
-
 export default class Home extends Component {
   constructor(props) {
     super(props);
@@ -30,7 +28,7 @@ export default class Home extends Component {
    */
   componentDidMount = async () => {
     const profilePics = this.importAll(
-      require.context("../images/emojis", false, /\.(png|jpe?g|svg)$/)
+      require.context("../SVG", false, /\.(png|jpe?g|svg)$/)
     );
     this.setState({ profilePics: profilePics });
     var elems = document.querySelectorAll(".dropdown-trigger");
@@ -79,6 +77,7 @@ export default class Home extends Component {
             teamMembers: teamInfo.members,
             teamPoints: teamPoints,
             timelineArr: timelineArr,
+            profilePic: teamInfo.userAnimal,
           });
         }
       });
@@ -87,7 +86,7 @@ export default class Home extends Component {
   importAll = (r) => {
     return r.keys().map(r);
   };
-  onClickBlacklistButton = (e) => {
+  onClickBlacklistButton = () => {
     if (!this.state.isInBlacklist) {
       this.setState({
         isInBlacklist: true,
@@ -133,7 +132,7 @@ export default class Home extends Component {
       </Button>
     );
   };
-  handleCheckIn = (e) => {
+  handleCheckIn = () => {
     this.setState({
       isCheckIn: !this.state.isCheckIn,
     });
@@ -204,7 +203,9 @@ export default class Home extends Component {
     return Math.round((num + Number.EPSILON) * 100) / 100;
   };
   getProfilePic = (profilePic) => {
-    console.log(profilePic);
+    if (profilePic == undefined) {
+      profilePic = "Predator";
+    }
     return require("../images/emojis/" + profilePic + ".svg");
   };
 
@@ -250,7 +251,11 @@ export default class Home extends Component {
             {/* <!-- Dropdown Structure --> */}
             <ul id="people" className="dropdown-content collection">
               {this.state.teamMembers.map((teamMember) => {
-                return <li className="collection-item">{teamMember}</li>;
+                return (
+                  <li key={teamMember} className="collection-item">
+                    {teamMember}
+                  </li>
+                );
               })}
             </ul>
           </div>
